create useOpenRecordsSearchPageInCommandMenu
