+++ packages/twenty-front/src/modules/command-menu/hooks/useCommandMenuHotKeys.ts
@@ -1,6 +1,7 @@
 import { COMMAND_MENU_COMPONENT_INSTANCE_ID } from '@/command-menu/constants/CommandMenuComponentInstanceId';
 import { useCommandMenu } from '@/command-menu/hooks/useCommandMenu';
 import { useCommandMenuHistory } from '@/command-menu/hooks/useCommandMenuHistory';
+import { useOpenRecordsSearchPageInCommandMenu } from '@/command-menu/hooks/useOpenRecordsSearchPageInCommandMenu';
 import { commandMenuPageState } from '@/command-menu/states/commandMenuPageState';
 import { commandMenuSearchState } from '@/command-menu/states/commandMenuSearchState';
 import { CommandMenuPages } from '@/command-menu/types/CommandMenuPages';
@@ -14,11 +15,9 @@ import { useRecoilValue } from 'recoil';
 import { Key } from 'ts-key-enum';
 
 export const useCommandMenuHotKeys = () => {
-  const {
-    openRecordsSearchPage,
-    toggleCommandMenu,
-    setGlobalCommandMenuContext,
-  } = useCommandMenu();
+  const { toggleCommandMenu, setGlobalCommandMenuContext } = useCommandMenu();
+
+  const { openRecordsSearchPage } = useOpenRecordsSearchPageInCommandMenu();
 
   const { goBackFromCommandMenu } = useCommandMenuHistory();
