+++ packages/twenty-front/src/modules/command-menu/hooks/useOpenRecordsSearchPageInCommandMenu.ts
@@ -1,0 +1,21 @@
+import { useCommandMenu } from '@/command-menu/hooks/useCommandMenu';
+import { CommandMenuPages } from '@/command-menu/types/CommandMenuPages';
+import { IconSearch } from 'twenty-ui';
+import { v4 } from 'uuid';
+
+export const useOpenRecordsSearchPageInCommandMenu = () => {
+  const { navigateCommandMenu } = useCommandMenu();
+
+  const openRecordsSearchPage = () => {
+    navigateCommandMenu({
+      page: CommandMenuPages.SearchRecords,
+      pageTitle: 'Search',
+      pageIcon: IconSearch,
+      pageId: v4(),
+    });
+  };
+
+  return {
+    openRecordsSearchPage,
+  };
+};
