+++ packages/twenty-front/src/modules/navigation/components/MainNavigationDrawerItems.tsx
@@ -2,7 +2,7 @@ import { useLocation } from 'react-router-dom';
 import { useRecoilState, useSetRecoilState } from 'recoil';
 import { IconSearch, IconSettings } from 'twenty-ui';
 
-import { useCommandMenu } from '@/command-menu/hooks/useCommandMenu';
+import { useOpenRecordsSearchPageInCommandMenu } from '@/command-menu/hooks/useOpenRecordsSearchPageInCommandMenu';
 import { CurrentWorkspaceMemberFavoritesFolders } from '@/favorites/components/CurrentWorkspaceMemberFavoritesFolders';
 import { WorkspaceFavorites } from '@/favorites/components/WorkspaceFavorites';
 import { NavigationDrawerOpenedSection } from '@/object-metadata/components/NavigationDrawerOpenedSection';
@@ -41,7 +41,7 @@ export const MainNavigationDrawerItems = () => {
 
   const { t } = useLingui();
 
-  const { openRecordsSearchPage } = useCommandMenu();
+  const { openRecordsSearchPage } = useOpenRecordsSearchPageInCommandMenu();
 
   return (
     <>
